requirements.txt
# Core API dependencies
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
python-dotenv==1.1.0
starlette==0.46.2

# ML/AI dependencies
torch==2.5.1+cu121
torchvision==0.20.1+cu121
transformers==4.51.3
accelerate==1.6.0
bitsandbytes==0.45.5
safetensors==0.5.3
tiktoken==0.9.0
huggingface-hub==0.30.2

# Image processing
pillow==11.0.0

# Utilities
numpy==2.1.2
tqdm==4.67.1
requests==2.32.3
pydantic==2.11.3